refactor: db 연결 객체 변경
