+++ src/main/java/repository/connection/ConnectDatabase.java
@@ -1,0 +1,7 @@
+package repository.connection;
+
+import java.sql.Connection;
+
+public interface ConnectDatabase {
+    Connection create();
+}
