+++ src/main/java/repository/connection/ConnectMysql.java
@@ -1,25 +1,23 @@
-package repository;
+package repository.connection;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public final class UserDao {
-
+public class ConnectMysql implements ConnectDatabase{
     private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
     private static final String DATABASE = "janggi"; // MySQL DATABASE 이름
     private static final String OPTION = "?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC";
     private static final String USERNAME = "user"; //  MySQL 서버 아이디
     private static final String PASSWORD = "password"; // MySQL 서버 비밀번호
 
-    public Connection getConnection() {
+    @Override
+    public Connection create() {
         // 드라이버 연결
         try {
             return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
         } catch (final SQLException e) {
-            System.err.println("DB 연결 오류:" + e.getMessage());
-            e.printStackTrace();
-            return null;
+            throw new RuntimeException("[ERROR] DB 연결 오류");
         }
     }
 }
